Create guess.py
It's a simple guessing game. It has to be run in the terminal. After you type the python file name, you have to choose numbers. the numbers you choose will then be turned into a start and finish for a range of numbers to choose from. The loop keeps running it until you guess it right.